backend/flaskBackend_revamped/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
nltk==3.5
numpy==1.19.5
pandas==1.2.3
Pillow==8.1.2
protobuf==3.15.3
PyJWT==2.0.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.1
SQLAlchemy==1.3.23
tensorflow==2.4.0
tensorflow-estimator==2.4.0
virtualenv==20.4.2
Werkzeug==1.0.1

